Add arguments to the script
Optimizing code.
Adding comments to code.
Ability to choose the language by an argument.
Remove remaining files is now optional, false by default.

+++ main.py
@@ -5,56 +5,69 @@ from pathlib import Path
 import shutil
 import argparse
 
+
 def main():
+    args = parse_arguments()
 
-    # get list of subtitles files in the desired language
-    all_files = set(glob.glob('./Subs/**/*.srt', recursive=True))
-    excluded_files = set(glob.glob('./Subs/**/*English*', recursive=True))
-    to_remove_files = list(all_files - excluded_files)
-
-    # remove files in another language
-    for file in to_remove_files:
-        try:
-            # os.remove(file)
-            print(str(file))
-        except OSError:
-            print("error")
-
-    # from all the subtitles in our language, discard the ones with only foreign dialogues or audio descriptive
     folders = glob.glob('./Subs/*')
     subs_to_move = []
-    for f in folders:
-        subs = glob.glob(f+"/*")
-        sizes = [os.path.getsize(f) for f in subs]
+    for f in folders:  # for each episode
+
+        subs = glob.glob(f + "/*" + args.language.capitalize() + "*.srt")  # get list of subtitles matching language
+        sizes = [os.path.getsize(f) for f in subs]  # get size of each subtitle file
         mean = np.mean(sizes)
 
         if len(sizes) > 2:
-            idx = (np.abs(np.asarray(sizes)-mean)).argmin()
+            # if there are more than 2 subtitles in the same language, the first is only foreign dialogues,
+            # the last is audio description for deaf persons.
+            # The ones in the middle are most likely to be a transcription of only the dialogues.
+            # By doing the closest to the mean, it is ensured to pick the right one
+            idx = (np.abs(np.asarray(sizes) - mean)).argmin()
             subs_to_move.append(subs[idx])
             continue
+
         if len(sizes) == 2:
+            # if there are only two files, there are to scenarios:
             std = np.std(sizes)
             cv = std / mean
             if cv > 0.5:
-                idx = (np.abs(np.asarray(sizes)-mean)).argmax()
+                # the first being foreign language transcription and normal transcription.
+                # This happens when the STD is high (one file has a small size and the other a big size).
+                # Picking the bigger size one.
+                idx = (np.abs(np.asarray(sizes) - mean)).argmax()
                 subs_to_move.append(subs[idx])
             else:
+                # the second being normal transcription and audio description for deaf person.
+                # both files have similar size, and thus a lower std. Picking the smaller size one
                 idx = (np.abs(np.asarray(sizes) - mean)).argmin()
                 subs_to_move.append(subs[idx])
             continue
+
         if len(sizes) == 1:
             subs_to_move.append(subs[0])
+            continue
+
 
     # rename and move files
     for sub in subs_to_move:
         parts = Path(sub).parts
         suffix = Path(sub).suffix
-        os.rename(sub, "./"+parts[1]+suffix)
+        shutil.copy(sub, "./" + parts[1] + suffix)
 
     # remove Subs folder and remaining files
-    shutil.rmtree('./Subs')
+    if args.remove_subs_dir:
+        shutil.rmtree('./Subs')
 
 
+def parse_arguments():
+    parser = argparse.ArgumentParser()
+    parser.add_argument('-l', '--language', default='English',
+                        help="Language for the subtitles you want e.g. --language=English")
+    parser.add_argument('-r', '--remove_subs_dir', default=False, action='store_true',
+                        help='use if you want the Subs folder to be deleted when the script finishes')
+    args = parser.parse_args()
+    return args
 
-if __name__ =="__main__":
+
+if __name__ == "__main__":
     main()
